evaluation/ImageBuilder/requirements.txt
aiohttp==3.8.4
aiosignal==1.2.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==22.2.0
biopython==1.79
centrosome==1.1.5
certifi==2023.5.7
charset-normalizer==2.0.12
cloudpickle==2.2.1
colorama
cycler==0.11.0
Cython==0.29.35
dataclasses==0.8
datasets==2.4.0
decorator==4.4.2
deprecation==2.1.0
dill==0.3.4
filelock==3.4.1
frozenlist==1.2.0
fsspec==2022.1.0
huggingface-hub==0.4.0
idna==3.4
idna-ssl==1.1.0
imageio==2.15.0
importlib-metadata==4.8.3
joblib
kiwisolver==1.3.1
matplotlib==3.3.4
multidict==5.2.0
multiprocess==0.70.12.2
networkx==2.5.1
numpy
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
pyarrow==6.0.1
pyparsing==3.0.9
pysam==0.21.0
python-dateutil
pytz
PyWavelets==1.1.1
PyYAML==6.0
requests==2.27.1
responses==0.17.0
scikit-image==0.17.2
scikit-learn
scipy==1.5.4
six
threadpoolctl
tifffile==2020.9.3
tqdm
typing_extensions==4.1.1
urllib3==1.26.16
xxhash==3.2.0
yarl==1.7.2
zipp==3.6.0
requests
bs4